requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
idna==3.6
pymongo==3.12.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.2.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
